fix: local date to date, list in list, math.round()

+++ src/main/java/com/github/moviereservationbe/repository/MainPage/movie/Movie.java
@@ -36,10 +36,10 @@ public class Movie {
     private Date releaseDate;
 
     @Column(name = "ticket_sales", nullable = false)
-    private Integer ticketSales;
+    private Double ticketSales;
 
     @Column(name = "score_avg", nullable = false)
-    private Integer scoreAvg;
+    private Double scoreAvg;
 
     @Column(name = "age_limit", nullable = false)
     private Integer ageLimit;
